--- Game State Stack ---
Game loop is properly added as a callback in window.

A function InitGame is called that addes a main menu to the
stack. This function does not currently work. I'm going to quickly
add main menu as just being an empty element, for sake of testing.

Next on my list is to implement a state frame factory.